Jenny Nguyen_ Assignment 4
Write a program that prompts the user for their monthly SDG&E bill for the last 5 months.

The program should find and output their average monthly SDG&E bill. If the average bill exceeds $100, the output should include a message indicating that too much Gas and Electricity is being used. If the average bill is at least $60 but no more than $100, the output should say that a typical amount of Gas and Electricity is being used. Finally, if the average bill is less than $60 the output should include a message congratulating the user for conserving Gas and Electricity.